+++ bugwarrior/services/activecollab2.py
@@ -204,11 +204,6 @@ class ActiveCollab2Service(Service):
                                           self.config.projects,
                                           self.config.target)
 
-    def get_owner(self, issue):
-        # TODO
-        raise NotImplementedError(
-            "This service has not implemented support for 'only_if_assigned'.")
-
     def issues(self):
         # Loop through each project
         start = time.time()
