+++ bugwarrior/services/base.py
@@ -106,24 +106,6 @@ class Service(abc.ABC):
                 final.append('@%s - %s' % (author, message))
         return final
 
-    def include(self, issue):
-        """ Return true if the issue in question should be included """
-        if self.config.only_if_assigned:
-            owner = self.get_owner(issue)
-            include_owners = [self.config.only_if_assigned]
-
-            if self.config.also_unassigned:
-                include_owners.append(None)
-
-            return owner in include_owners
-
-        return True
-
-    @abc.abstractmethod
-    def get_owner(self, issue):
-        """ Override this for filtering on tickets """
-        raise NotImplementedError()
-
     @abc.abstractmethod
     def issues(self):
         """ Returns a list of Issue instances representing issues from a remote service.
