+++ bugwarrior/services/gmail.py
@@ -204,10 +204,6 @@ class GmailService(Service):
         issue_url = issue.extra['url']
         return self.build_annotations([(sender, subj)], issue_url)
 
-    def get_owner(self, issue):
-        raise NotImplementedError(
-            "This service has not implemented support for 'only_if_assigned'.")
-
     def issues(self):
         labels = self.get_labels()
         for thread in self.get_threads():
