+++ bugwarrior/services/pagure.py
@@ -154,6 +154,19 @@ class PagureService(Service):
         if issue[1]['assignee']:
             return issue[1]['assignee']['name']
 
+    def include(self, issue):
+        """ Return true if the issue in question should be included """
+        if self.config.only_if_assigned:
+            owner = self.get_owner(issue)
+            include_owners = [self.config.only_if_assigned]
+
+            if self.config.also_unassigned:
+                include_owners.append(None)
+
+            return owner in include_owners
+
+        return True
+
     def filter_repos(self, repo):
         if repo in self.config.exclude_repos:
             return False
